# [BEAM-4727] Improve metric lookup overhead for predeclared metrics